+++ test/bigtest/interactors/usage-consolidation-filters.js
@@ -4,15 +4,16 @@ import {
   collection,
   text,
 } from '@bigtest/interactor';
-import SelectInteractor from '@folio/stripes-components/lib/Select/tests/interactor';
+
+import Select from './select';
 
 @interactor class UsageConsolidationFilters {
-  yearDropdown = new SelectInteractor('[data-test-usage-consolidation-year-filter]');
+  yearDropdown = new Select('[data-test-usage-consolidation-year-filter]');
   yearDropdownOptions = collection('[data-test-usage-consolidation-year-filter] option', {
     value: text(),
   });
 
-  platformTypeDropdown = new SelectInteractor('[data-test-usage-consolidation-platform-type-filter]');
+  platformTypeDropdown = new Select('[data-test-usage-consolidation-platform-type-filter]');
   clickView = clickable('[data-test-usage-consolidation-view-button]');
 }